adds working movie database

+++ db/migrate/20190729205816_create_movies.rb
@@ -6,7 +6,7 @@ class CreateMovies < ActiveRecord::Migration[5.2]
       t.string :genre
       t.string :overview
       t.string :poster_path
-      t.integer :rating
+      t.float :rating
 
       t.timestamps
     end
